requirements.txt
Flask==2.3.3
pandas==2.0.3
Werkzeug==2.3.7
openai>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0 